feat(freertos): increase idle task stack 192->500
Prior to this commit, the idle task stack size was the default of 192
bytes.  However, this caused a stack overflow when the serial packet
handler processed a complete packet request.

This commit increases the idle task stack size from 192 bytes to 500 bytes
to prevent the serial packet handler from causing a stack overflow.