feat(task/motor): add at lowest priority
As of this commit, the serial port is listening in the idle task (which
has the *lowest* priority).

This commit adds a task to process requests to move the motor.  The motor
task is set to the lowest priority to allow the idle task to run even if
the motor task was frequently busy.

Notes
-----

According to [here][1], the largest useful delay value for
`delayMicroseconds` is 16383.

As a workaround within the motor task, divide specified microsecond delay
by ~1000 (actually by 1024, since this can be done using a shift operation
instead of actual division) and use the millisecond delay function for the
result and the microsecond delay function for the remainder.

[1]: https://stackoverflow.com/questions/34532941/in-arduino-is-there-a-maximum-delay-time-when-using-the-fuctiondelay/34533207#34533207